+++ kymatio/scattering2d/__init__.py
@@ -1,4 +1,4 @@
 __all__ = ['Scattering2D']
 
+from .frontend import Scattering2D
 
-from .scattering2d import Scattering2D
