--- kymatio/scattering2d/backend/backend_skcuda.py
@@ -1,347 +1,0 @@
-# Authors: Edouard Oyallon, Sergey Zagoruyko
-
-from collections import namedtuple
-import torch
-from skcuda import cublas
-import cupy
-from string import Template
-from torch.nn import ReflectionPad2d
-import numpy as np
-
-NAME = 'skcuda'
-
-@cupy.util.memoize(for_each_device=True)
-def load_kernel(kernel_name, code, **kwargs):
-    code = Template(code).substitute(**kwargs)
-    kernel_code = cupy.cuda.compile_with_cache(code)
-    return kernel_code.get_function(kernel_name)
-
-
-Stream = namedtuple('Stream', ['ptr'])
-
-
-def convert_filters(bank):
-    for c, psi in enumerate(bank):
-        if isinstance(psi, np.ndarray):
-            bank[c] = torch.from_numpy(bank[c])
-        else:
-            for k, v in psi.items():
-                if isinstance(k, int):
-                    bank[c][k] = torch.from_numpy(v)
-    return bank
-
-
-
-def getDtype(t):
-    if isinstance(t, torch.cuda.FloatTensor):
-        return 'float'
-    elif isinstance(t, torch.cuda.DoubleTensor):
-        return 'double'
-
-
-def iscomplex(input):
-    return input.size(-1) == 2
-
-
-def isreal(input):
-    return input.size(-1) == 1
-
-
-class Pad(object):
-    def __init__(self, pad_size, input_size, pre_pad=False):
-        """
-            Padding which allows to simultaneously pad in a reflection fashion
-            and map to complex.
-
-            Parameters
-            ----------
-            pad_size : list of 4 integers
-                size of padding to apply [top, bottom, left, right].
-            input_size : list of 2 integers
-                size of the original signal [height, width].
-            pre_pad : boolean
-                if set to true, then there is no padding, one simply adds the imaginary part.
-        """
-        self.pre_pad = pre_pad
-        self.pad_size = pad_size
-        self.input_size = input_size
-
-        self.build()
-
-    def build(self):
-        pad_size_tmp = list(self.pad_size)
-
-        # This allow to handle the case where the padding is equal to the image size
-        if pad_size_tmp[0] == self.input_size[0]:
-            pad_size_tmp[0] -= 1
-            pad_size_tmp[1] -= 1
-        if pad_size_tmp[2] == self.input_size[1]:
-            pad_size_tmp[2] -= 1
-            pad_size_tmp[3] -= 1
-        # Pytorch expects its padding as [left, right, top, bottom]
-        self.padding_module = ReflectionPad2d([pad_size_tmp[2], pad_size_tmp[3],
-                                               pad_size_tmp[0], pad_size_tmp[1]])
-
-    def __call__(self, x):
-        if not self.pre_pad:
-            x = self.padding_module(x)
-            if self.pad_size[0] == self.input_size[0]:
-                x = torch.cat([x[:, :, 1, :].unsqueeze(2), x, x[:, :, x.size(2) - 2, :].unsqueeze(2)], 2)
-            if self.pad_size[2] == self.input_size[1]:
-                x = torch.cat([x[:, :, :, 1].unsqueeze(3), x, x[:, :, :, x.size(3) - 2].unsqueeze(3)], 3)
-
-        output = x.new_zeros(x.shape + (2,))
-        output[...,0] = x
-        return output
-
-def unpad(in_):
-    """
-        Slices the input tensor at indices between 1::-1
-
-        Parameters
-        ----------
-        in_ : tensor_like
-            input tensor
-
-        Returns
-        -------
-        in_[..., 1:-1, 1:-1]
-    """
-    return in_[..., 1:-1, 1:-1]
-
-class SubsampleFourier(object):
-    """
-        Subsampling of a 2D image performed in the Fourier domain
-        Subsampling in the spatial domain amounts to periodization
-        in the Fourier domain, hence the formula.
-
-        Parameters
-        ----------
-        x : tensor_like
-            input tensor with at least 5 dimensions, the last being the real
-             and imaginary parts.
-            Ideally, the last dimension should be a power of 2 to avoid errors.
-        k : int
-            integer such that x is subsampled by 2**k along the spatial variables.
-
-        Returns
-        -------
-        res : tensor_like
-            tensor such that its fourier transform is the Fourier
-            transform of a subsampled version of x, i.e. in
-            FFT^{-1}(res)[u1, u2] = FFT^{-1}(x)[u1 * (2**k), u2 * (2**k)]
-    """
-    def __init__(self):
-        self.block = (32, 32, 1)
-
-    def GET_BLOCKS(self, N, threads):
-        return (N + threads - 1) // threads
-
-    def __call__(self, input, k):
-        if not input.is_cuda:
-            raise RuntimeError('Use the torch backend for cpu tensors!')
-
-        out = input.new(input.size(0), input.size(1), input.size(2) // k, input.size(3) // k, 2)
-
-        if not iscomplex(input):
-            raise (TypeError('The input and outputs should be complex'))
-
-        input = input.contiguous()
-
-        kernel = '''
-        #define NW ${W} / ${k}
-        #define NH ${H} / ${k}
-        extern "C"
-        __global__ void periodize(const ${Dtype}2 *input, ${Dtype}2 *output)
-        {
-          int tx = blockIdx.x * blockDim.x + threadIdx.x;
-          int ty = blockIdx.y * blockDim.y + threadIdx.y;
-          int tz = blockIdx.z * blockDim.z + threadIdx.z;
-          if(tx >= NW || ty >= NH || tz >= ${B})
-            return;
-          input += tz * ${H} * ${W} + ty * ${W} + tx;
-          ${Dtype}2 res = make_${Dtype}2(0.f, 0.f);
-          for (int j=0; j<${k}; ++j)
-            for (int i=0; i<${k}; ++i)
-            {
-              const ${Dtype}2 &c = input[j * NH * ${W} + i * NW];
-              res.x += c.x;
-              res.y += c.y;
-            }
-          res.x /= ${k} * ${k};
-          res.y /= ${k} * ${k};
-          output[tz * NH * NW + ty * NW + tx] = res;
-        }
-        '''
-        B = input.nelement() // (2*input.size(-2) * input.size(-3))
-        W = input.size(-2)
-        H = input.size(-3)
-        k = input.size(-2) // out.size(-2)
-        periodize = load_kernel('periodize', kernel, B=B, H=H, W=W, k=k, Dtype=getDtype(input))
-        grid = (self.GET_BLOCKS(out.size(-3), self.block[0]),
-                self.GET_BLOCKS(out.size(-2), self.block[1]),
-                self.GET_BLOCKS(out.nelement() // (2*out.size(-2) * out.size(-3)), self.block[2]))
-        periodize(grid=grid, block=self.block, args=[input.data_ptr(), out.data_ptr()],
-                  stream=Stream(ptr=torch.cuda.current_stream().cuda_stream))
-        return out
-
-
-class Modulus(object):
-    """
-        This class implements a modulus transform for complex numbers.
-
-        Usage
-        -----
-        modulus = Modulus()
-        x_mod = modulus(x)
-
-        Parameters
-        ---------
-        x: input tensor, with last dimension = 2 for complex numbers
-
-        Returns
-        -------
-        output: a tensor with imaginary part set to 0, real part set equal to
-        the modulus of x.
-    """
-    def __init__(self):
-        self.CUDA_NUM_THREADS = 1024
-
-    def GET_BLOCKS(self, N):
-        return (N + self.CUDA_NUM_THREADS - 1) // self.CUDA_NUM_THREADS
-
-    def __call__(self, input):
-        if not input.is_cuda:
-            raise RuntimeError('Use the torch backend for cpu tensors!')
-
-        out = input.new(input.size())
-        input = input.contiguous()
-
-        if not iscomplex(input):
-            raise TypeError('The input and outputs should be complex')
-
-        kernel = """
-        extern "C"
-        __global__ void abs_complex_value(const ${Dtype} * x, ${Dtype}2 * z, int n)
-        {
-            int i = blockIdx.x * blockDim.x + threadIdx.x;
-        if (i >= n)
-            return;
-        z[i] = make_${Dtype}2(normf(2, x + 2*i), 0);
-
-        }
-        """
-        fabs = load_kernel('abs_complex_value', kernel, Dtype=getDtype(input))
-        fabs(grid=(self.GET_BLOCKS(int(out.nelement())//2), 1, 1),
-             block=(self.CUDA_NUM_THREADS, 1, 1),
-             args=[input.data_ptr(), out.data_ptr(), out.numel() // 2],
-             stream=Stream(ptr=torch.cuda.current_stream().cuda_stream))
-        return out
-
-
-
-
-def fft(input, direction='C2C', inverse=False):
-    """
-        Interface with torch FFT routines for 2D signals.
-
-        Example
-        -------
-        x = torch.randn(128, 32, 32, 2)
-        x_fft = fft(x, inverse=True)
-
-        Parameters
-        ----------
-        input : tensor
-            complex input for the FFT
-        direction : string
-            'C2R' for complex to real, 'C2C' for complex to complex
-        inverse : bool
-            True for computing the inverse FFT.
-            NB : if direction is equal to 'C2R', then the transform
-            is automatically inverse.
-    """
-    if direction == 'C2R':
-        inverse = True
-
-    if not iscomplex(input):
-        raise(TypeError('The input should be complex (e.g. last dimension is 2)'))
-
-    if (not input.is_contiguous()):
-        raise (RuntimeError('Tensors must be contiguous!'))
-
-    if direction == 'C2R':
-        output = torch.irfft(input, 2, normalized=False, onesided=False)*input.size(-2)*input.size(-3)
-    elif direction == 'C2C':
-        if inverse:
-            output = torch.ifft(input, 2, normalized=False)*input.size(-2)*input.size(-3)
-        else:
-            output = torch.fft(input, 2, normalized=False)
-
-    return output
-
-
-
-
-def cdgmm(A, B, inplace=False):
-    """
-        Complex pointwise multiplication between (batched) tensor A and tensor B.
-
-        Parameters
-        ----------
-        A : tensor
-            A is a complex tensor of size (B, C, M, N, 2)
-        B : tensor
-            B is a complex tensor of size (M, N, 2) or real tensor of (M, N, 1)
-        inplace : boolean, optional
-            if set to True, all the operations are performed inplace
-
-        Returns
-        -------
-        C : tensor
-            output tensor of size (B, C, M, N, 2) such that:
-            C[b, c, m, n, :] = A[b, c, m, n, :] * B[m, n, :]
-    """
-    if not iscomplex(A):
-        raise TypeError('The input must be complex, indicated by a last '
-                        'dimension of size 2')
-
-    if B.ndimension() != 3:
-        raise RuntimeError('The filter must be a 3-tensor, with a last '
-                           'dimension of size 1 or 2 to indicate it is real '
-                           'or complex, respectively')
-
-    if not iscomplex(B) and not isreal(B):
-        raise TypeError('The filter must be complex or real, indicated by a '
-                        'last dimension of size 2 or 1, respectively')
-
-    if A.size()[-3:-1] != B.size()[-3:-1]:
-        raise RuntimeError('The filters are not compatible for multiplication!')
-
-    if A.dtype is not B.dtype:
-        raise RuntimeError('A and B must be of the same dtype')
-
-    if A.device.type != B.device.type:
-        raise RuntimeError('A and B must be of the same device type')
-
-    if A.device.type == 'cuda':
-        if A.device.index != B.device.index:
-            raise RuntimeError('A and B must be on the same GPU!')
-
-    if isreal(B):
-        if inplace:
-            return A.mul_(B)
-        else:
-            return A * B
-    else:
-        A, B = A.contiguous(), B.contiguous()
-        C = A.new(A.size()) if not inplace else A
-        m, n = B.nelement() // 2, A.nelement() // B.nelement()
-        lda = m
-        ldc = m
-        incx = 1
-        handle = torch.cuda.current_blas_handle()
-        stream = torch.cuda.current_stream()._as_parameter_
-        cublas.cublasSetStream(handle, stream)
-        cublas.cublasCdgmm(handle, 'l', m, n, A.data_ptr(), lda, B.data_ptr(), incx, C.data_ptr(), ldc)
-        return C
